requirements.txt
assemblyai==0.35.1
yt-dlp==2024.11.18
TTS==0.21.1
moviepy==1.0.3
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
deep-translator==1.11.4
httpx==0.27.2
websockets==14.1
torch==2.1.0
numpy==1.24.0
ffmpeg-python==0.2.0
streamlit==1.24.0
protobuf==3.20.0
customtkinter==5.2.1
pillow==9.5.0
darkdetect==0.8.0
